Moving utils and constants to their new logical places
