+++ common/constants/colors.ts
@@ -24,3 +24,30 @@ export const COLORS = {
     background: '#F6F6F6',
   },
 };
+
+// Colors for charts
+export const CHART_COLORS = {
+  GREY: '#1c1c1c',
+  GREEN: '#64b96a',
+  YELLOW: '#f5ed00',
+  RED: '#c33149',
+  PURPLE: '#bb5cc1',
+  FILL: '#bfc8d680',
+  FILL_HOURLY: '#88aee680',
+};
+
+export const LINE_COLORS = {
+  RED: '#da291c',
+  ORANGE: '#ed8b00',
+  BLUE: '#003da5',
+  GREEN: '#00834d',
+  BUS: '#ffc72c',
+};
+
+export const colorsForLine: Record<string, string> = {
+  red: LINE_COLORS.RED,
+  orange: LINE_COLORS.ORANGE,
+  blue: LINE_COLORS.BLUE,
+  green: LINE_COLORS.GREEN,
+  bus: LINE_COLORS.BUS,
+};
