+++ common/constants/lines.ts
@@ -38,3 +38,11 @@ export const LINE_OBJECTS: LineObject = {
     color: COLORS.mbta.bus,
   },
 };
+
+export const lineTabs = [
+  { name: 'Red Line', current: false, id: 'red' },
+  { name: 'Blue Line', current: false, id: 'blue' },
+  { name: 'Orange Line', current: false, id: 'orange' },
+  { name: 'Green Line', current: false, id: 'green' },
+  { name: 'Bus', current: false, id: 'bus' },
+];
