+++ common/utils/date.ts
@@ -17,3 +17,11 @@ export const prettyDate = (dateString: string, withDow: boolean) => {
     options
   );
 };
+
+export const getCurrentDate = (): string => {
+  const isoDate = new Date();
+  const offset = isoDate.getTimezoneOffset();
+  const localDate = new Date(isoDate.valueOf() - offset * 60 * 1000);
+  const maxDate = localDate.toISOString().split('T')[0];
+  return maxDate;
+};
