+++ modules/dashboard/DataPageHeader.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import Datepicker from 'react-tailwindcss-datepicker';
 import type { DateValueType } from 'react-tailwindcss-datepicker/dist/types';
 import { LINE_OBJECTS } from '../../common/constants/lines';
-import { getCurrentDate } from '../../utils/date';
+import { getCurrentDate } from '../../common/utils/date';
 import { Button } from '../../common/components/inputs/Button';
 import { useDelimitatedRoute } from '../../common/utils/router';
 import { useBreakpoint } from '../../common/hooks/useBreakpoint';
