--- utils/date.ts
@@ -1,7 +1,0 @@
-export const getCurrentDate = (): string => {
-  const isoDate = new Date();
-  const offset = isoDate.getTimezoneOffset();
-  const localDate = new Date(isoDate.valueOf() - offset * 60 * 1000);
-  const maxDate = localDate.toISOString().split('T')[0];
-  return maxDate;
-};
