--- utils/lines.ts
@@ -1,7 +1,0 @@
-export const lineTabs = [
-  { name: 'Red Line', current: false, id: 'red' },
-  { name: 'Blue Line', current: false, id: 'blue' },
-  { name: 'Orange Line', current: false, id: 'orange' },
-  { name: 'Green Line', current: false, id: 'green' },
-  { name: 'Bus', current: false, id: 'bus' },
-];
